refactor: instances management
