+++ src/process/index.js
@@ -2,6 +2,7 @@ import { app } from "electron";
 import electronUpdater from "electron-updater";
 import { MainWindow } from "./window.js";
 
+await import("./instance.js");
 await import("./navigation.js");
 
 app.enableSandbox();
