+++ src/types/ipc.ts
@@ -3,6 +3,11 @@ import { Settings } from "../process/settings.js";
 
 export type Api = {
 	send: (channel: string, data?: unknown) => void;
+	instance: {
+		register: (instance: Partial<Settings["instances"][number]>) => void;
+		remove: (id: string) => void;
+		setDefault: (id: string) => void;
+	};
 	setTheme: (themeId: NativeTheme["themeSource"]) => void;
 	getSetting: <S extends keyof Settings>(setting: S) => Promise<Settings[S]>;
 	setSetting: <S extends keyof Settings>(
